pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wednesday-mcp"
version = "1.0.0"
description = "MCP Server for Wednesday WhatsApp Assistant - AI-powered WhatsApp assistant with calendar, email, tasks, and more"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Wednesday Assistant" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "bytez>=3.0.0",
    "google-api-python-client>=2.115.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.2.0",
    "spotipy>=2.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
full = [
    "google-cloud-speech",
    "google-cloud-texttospeech",
    "firebase-admin",
]

[project.scripts]
wednesday-mcp = "mcp_server.server:main"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
